carythma/requirements.txt
argcomplete==3.1.1
asgiref==3.7.2
boto3==1.28.49
botocore==1.31.49
certifi==2023.7.22
cfn-flip==1.3.0
charset-normalizer==3.2.0
click==8.1.7
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
durationpy==0.5
hjson==3.1.0
idna==3.4
jmespath==1.0.1
joblib==1.3.2
kappa==0.6.0
Markdown==3.4.4
MarkupSafe==2.1.3
numpy==1.25.1
placebo==0.9.0
psycopg==3.1.10
psycopg-binary==3.1.10
psycopg2==2.9.7
psycopg2-binary==2.9.7
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.1
troposphere==4.4.1
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.7
zappa==0.58.0
